pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-login>=0.6.3",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "oauthlib>=3.2.2",
    "flask-mail>=0.10.0",
    "pillow>=11.2.1",
    "requests>=2.32.4",
    "werkzeug>=3.1.3",
    "python-dotenv>=1.1.0",
    "sqlalchemy>=2.0.41",
    "flask-migrate>=4.1.0",
]
